requirements.txt
# Core OCR Framework
python-doctr[torch]>=0.9.0

# Azure Services
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-mgmt-containerinstance>=10.1.0
azure-mgmt-containerregistry>=10.1.0

# Label Studio Integration
label-studio-sdk>=1.0.0

# Training & ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
mlflow>=2.8.0
pyyaml>=6.0

# Data Processing
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# API & Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: DVC for data versioning
# dvc[azure]>=3.0.0

